refactor: 풀이 수정/삭제 컴포넌트 분리
