+++ frontend/src/components/SolutionDetail/index.tsx
@@ -2,13 +2,8 @@ import * as S from './SolutionSection.styled';
 import type { Solution } from '@/types/solution';
 import Button from '@/components/common/Button/Button';
 import SolutionDetailHeader from './SolutionDetailHeader';
-import { useSolutionDelete } from '@/hooks/useSolutionDelete';
-import { useNavigate } from 'react-router-dom';
-import { ROUTES } from '@/constants/routes';
-import Modal from '@/components/common/Modal/Modal';
-import useModal from '@/hooks/useModal';
 import useUserInfo from '@/hooks/useUserInfo';
-import ConfirmButtons from '@/components/ModalContent/ConfirmButtons';
+import SolutionBottom from './SolutionBottom';
 
 interface SolutionDetailProps {
   solution: Solution;
@@ -17,20 +12,8 @@ interface SolutionDetailProps {
 export default function SolutionSection({ solution }: SolutionDetailProps) {
   const { id: solutionId, description, url, mission } = solution;
   // 수정, 삭제 잘 되는지 dev에서 확인 필요합니다. @프룬
-  const { solutionDeleteMutation } = useSolutionDelete();
-
-  const navigate = useNavigate();
-  const handleNavigateToModifySolution = () => {
-    navigate(`${ROUTES.submitSolution}/${mission.id}?solutionId=${solutionId}`);
-  };
 
   const { data: userInfo } = useUserInfo();
-  const { isModalOpen, handleModalClose, handleModalOpen } = useModal();
-
-  const handleSolutionDelete = () => {
-    solutionDeleteMutation(solutionId);
-    handleModalClose();
-  };
 
   return (
     <section>
@@ -45,23 +28,9 @@ export default function SolutionSection({ solution }: SolutionDetailProps) {
         </S.CodeViewButtonLink>
       </S.CodeViewButtonWrapper>
       <S.SolutionDescription>{description}</S.SolutionDescription>
-      {userInfo?.id === solution.member.id && (
-        <S.SolutionDescriptionBottom>
-          <Button variant="defaultText" onClick={() => handleNavigateToModifySolution()}>
-            수정
-          </Button>
-          <Button variant="defaultText" onClick={handleModalOpen}>
-            삭제
-          </Button>
-        </S.SolutionDescriptionBottom>
+      {userInfo?.id !== solution.member.id && (
+        <SolutionBottom missionId={mission.id} solutionId={solutionId} />
       )}
-
-      <Modal isModalOpen={isModalOpen}>
-        <ConfirmButtons
-          handleModalClose={handleModalClose}
-          handleSolutionDelete={handleSolutionDelete}
-        />
-      </Modal>
     </section>
   );
 }
